moved classic model solver to separate class
